backend/requirements.txt
Django>=5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
requests>=2.31
psycopg2-binary>=2.9
python-dotenv>=1.0
dj-database-url>=2.1
